Change skeleton2

+++ skeleton2/skeleton/project/hardware/heavy_hardware.py
@@ -3,7 +3,7 @@ from project.hardware.hardware import Hardware
 
 class HeavyHardware(Hardware):
     def __init__(self, name: str, capacity: int, memory: int):
-        super().__init__(name, 'Heavy', capacity=capacity * 2, memory=memory * 0.75)
+        super().__init__(name, 'Heavy', capacity=capacity * 2, memory=int(memory * 0.75))
 
 
 if __name__ == '__main__':
